pyproject.toml
[project]
name = "agent-mcp"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "fastapi-mcp>=0.4.0",
    "fastapi>=0.116.1",
    "langchain>=0.3.27",
    "langgraph>=0.6.4",
    "langchain-mcp-adapters>=0.1.9",
    "langchain-openai>=0.3.29",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "fastmcp>=2.11.2",
]

[project.scripts]
api = "moov_api.__main__:main"

[tool.uv]
package = true

[dependency-groups]
dev = [
    "fastapi[standard]>=0.116.1",
    "pyright>=1.1.403",
]

[tool.ruff]
include = ["src/**/*.py", "tests/**/*.py"]

[tool.ruff.lint]
preview = true
extend-select = ["S", "I", "FBT", "A", "COM", "DTZ", "ISC", "LOG", "G", "PIE", "T20", "PT", "RET", "SIM", "TID", "PTH", "UP", "DOC"]
ignore = ["COM812", "ISC001"]

[tool.ruff.lint.per-file-ignores]
"**/tests/**" = ["S"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"src".msg = "Do not use src imports. Imports are relative to the src directory and should not include the 'src.' prefix."

[tool.ruff.format]
docstring-code-format = true

[tool.mypy]
strict = true
warn_unused_ignores = true
warn_return_any = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
